Moved IotHubCommander
