+++ tools/IotHubCommander.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotHubCommander", "IotHubCommander\IotHubCommander.csproj", "{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}"
 EndProject
