+++ tools/IotHubCommander/IotHubCommander.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotHubCommander", "IotHubCommander\IotHubCommander.csproj", "{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AABA282A-48A5-4C6D-B47F-7AE9E1293B7F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		EnterpriseLibraryConfigurationToolBinariesPathV6 = packages\EnterpriseLibrary.TransientFaultHandling.6.0.1304.0\lib\portable-net45+win+wp8
+	EndGlobalSection
+EndGlobal
