+++ tools/IotHubCommander/IotHubCommander/JsonTemplate.txt
@@ -1,0 +1,5 @@
+﻿{
+   "Temp" : @1,
+   "Humidity": @2,
+   "Counter": @3 
+}
