+++ tools/IotHubCommander/IotHubCommander/TextData2.csv
@@ -1,0 +1,14 @@
+﻿29.00;1987.12;3;1;
+31.21;1981.11;5;1;
+29.00;1987.12;3;0;
+31.21;1981.11;5;0;
+29.00;1987.12;3;0;
+31.21;1981.11;5;0;
+29.00;1987.12;3;0;
+31.21;1981.11;5;1;
+29.00;1987.12;3;0;
+31.21;1981.11;5;1;
+29.00;1987.12;3;0;
+31.21;1981.11;5;0;
+29.00;1987.12;3;0;
+31.21;1981.11;5;1;
