+++ tools/IotHubCommander/IotHubCommander/packages.config
@@ -1,0 +1,29 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="DotNetty.Buffers-signed" version="0.3.2" targetFramework="net452" />
+  <package id="DotNetty.Codecs.Mqtt-signed" version="0.3.2" targetFramework="net452" />
+  <package id="DotNetty.Codecs-signed" version="0.3.2" targetFramework="net452" />
+  <package id="DotNetty.Common-signed" version="0.3.2" targetFramework="net452" />
+  <package id="DotNetty.Handlers-signed" version="0.3.2" targetFramework="net452" />
+  <package id="DotNetty.Transport-signed" version="0.3.2" targetFramework="net452" />
+  <package id="EnterpriseLibrary.TransientFaultHandling" version="6.0.1304.0" targetFramework="net452" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net452" />
+  <package id="Microsoft.Azure.Amqp" version="1.1.5" targetFramework="net452" />
+  <package id="Microsoft.Azure.Devices" version="1.0.15" targetFramework="net452" />
+  <package id="Microsoft.Azure.Devices.Client" version="1.0.20" targetFramework="net452" />
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net452" />
+  <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.Framework.Configuration" version="1.0.0-beta8" targetFramework="net452" />
+  <package id="Microsoft.Framework.Configuration.Abstractions" version="1.0.0-beta8" targetFramework="net452" />
+  <package id="Microsoft.Framework.Configuration.CommandLine" version="1.0.0-beta8" targetFramework="net452" />
+  <package id="Mono.Security" version="3.2.3.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net452" />
+  <package id="PCLCrypto" version="1.0.86" targetFramework="net452" />
+  <package id="System.Spatial" version="5.6.4" targetFramework="net452" />
+  <package id="Validation" version="2.2.8" targetFramework="net452" />
+  <package id="WindowsAzure.ServiceBus" version="3.4.0" targetFramework="net452" />
+  <package id="WindowsAzure.Storage" version="7.0.0" targetFramework="net452" />
+</packages>
