--- tools/IotHubCommander/JsonTemplate.txt
@@ -1,5 +1,0 @@
-﻿{
-   "Temp" : @1,
-   "Humidity": @2,
-   "Counter": @3 
-}
