--- tools/IotHubCommander/TestData.csv
@@ -1,14 +1,0 @@
-﻿29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5	
-29.00	1987.12	3
-31.21	1981.11	5
